==> R/01_load.qmd <==
---
title: "Gene Expression Analysis"
format:
  html:
    embed-resources: true
editor: visual
---

## Data Origin

The gene expression data used in this project is available in the Gene Expression Omnibus (GEO) under the accession number GSE50834. The data originates from the following paper:

Dawany N, Showe LC, Kossenkov AV, Chang C et al. Identification of a 251 gene expression signature that can accurately detect M. tuberculosis in patients with and without HIV co-infection. PLoS One 2014;9(2):e89925. PMID: 24587128

## Summary

The data comes from total RNA isolated from peripheral blood mononuclear cell (PBMC) from a South African cohort. Microarry analysis was performed to compare gene expression in patients either infected with HIV or co-infected with HIV and tuberculosis (TB). Diagnosis of TB, especially in the presence of an HIV co-infection, can be challenging when using conventional diagnostic methods. Therefore, the data was used to identify a TB-specific gene signature.

## Setup

Loading required packages.

```{r}
#| label: 01_Load packages
#| message: false
library("tidyverse")
library("stringr")
library("GEOquery")
```

Note: Instruction of how to install the package GEOquery can be found here *https://bioconductor.org/packages/release/bioc/html/GEOquery.html*

## Load Data

This document will load the data from Gene Expression Omnibus. Furthermore, the file will export the data as a tsv-file. In the following, the data directory is defined along with the location of the original data.

```{r}
#| label: 01_Set up
dest_dir <- "../data/"
raw_dir <- str_c(dest_dir, "_raw/")
expr_file <- "GSE50834_non-normalized.txt.gz"
meta_file <- "GSE50834_series_matrix.txt.gz"
geo <- "GSE50834"

if( !dir.exists(raw_dir) ){
  dir.create(path = raw_dir,
             recursive = TRUE)
}
```

The gene expression data, containing patient ID's and gene expression, is downloaded and stored in the raw data directory using the GEOquery library. GEOquery can download data sets from the NCBI Gene Expression Omnibus (GEO), using a unique accession number.

```{r}
#| label: 01_Load expr-data
#| message: false
if( !file.exists(str_c(raw_dir, expr_file)) ){
  getGEOSuppFiles(GEO = geo,
                  makeDirectory = FALSE,
                  baseDir = raw_dir,
                  fetch_files = TRUE,
                  filter_regex = "txt")
}

dat_expr <- read.table(file = str_c(raw_dir, expr_file),
                         sep = "\t",
                         header = TRUE)
```

The meta data, containing patient information, is downloaded and stored in the raw data directory.

```{r}
#| label: 01_Load meta-data
#| message: false
if (!file.exists(str_c(raw_dir, meta_file))){
  dat_meta <- getGEO(GEO = geo)
  dat_meta <- pData(phenoData(dat_meta[[1]]))
  write.table(dat_meta, file = str_c(raw_dir, meta_file), 
             sep = "\t")
  dat_meta <- read.table(str_c(raw_dir, meta_file), 
                        sep = "\t",
                        header = TRUE)
} else {
  dat_meta <- read.table(str_c(raw_dir, meta_file), 
                          sep = "\t",
                          header = TRUE)
}
```

## Write Data to Files

The gene expression data and the meta data are saved as .tsv-files

```{r}
#| label: 01_Write tsv-files
dat_expr |> 
  write_tsv(file = str_c(dest_dir, "01_dat_load_expr.tsv"))

dat_meta |> 
  write_tsv(file = str_c(dest_dir, "01_dat_load_meta.tsv"))
```
