==> R/03_describe.qmd <==
---
title: "Gene Expression Analysis"
format:
  html:
    embed-resources: true
editor: visual
---

# Describe data

## Setup

Loading required packages

```{r}
#| label: 03_Load packages
#| message: false
library("tidyverse")
library("stringr")
library("table1")
```

Loading the cleaned data.

```{r}
#| label: 03_Load data
#| message: false
dat_clean <- read_tsv("../data/02_dat_clean.tsv")
```

## Data Description

Define setting for bar plot to illustrate the number of males and females for each patient group.

```{r}
#| label: 03_Define settings for plot
female_sign = intToUtf8(9792)
male_sign = intToUtf8(9794)
sign_size = 10
bar_width = 0.75

groups <- dat_clean |> 
  group_by(gender,
           disease_state) |> 
  summarize(n = n())
```

Create plot.

```{r}
#| label: 03_Bar plot
barplot_patient_group <- ggplot(data = dat_clean,
                                mapping = aes(x = disease_state,
                                              fill = disease_state,
                                              alpha = gender)) +
  geom_bar(width = bar_width,
           position = "dodge") +
  scale_fill_manual(
    values = c("HIV" = "royalblue", 
               "HIV/TB" = "tomato") ) +
  scale_alpha_manual(values = c("Male" = 0.6,
                                "Female" = 0.9)) +
  annotate("text", 
           x = 1:2 - bar_width/4, 
           y = groups$n[1:2] + 2, 
           label = female_sign, 
           size = sign_size) +
  annotate("text", 
           x = 1:2 + bar_width/4,
           y = groups$n[3:4] + 2, 
           label = male_sign, 
           size = sign_size) +
  coord_cartesian(ylim =  c(0, 20)) +
  theme_minimal(base_size = 16) + 
  labs(
    x = "Disease state",
    y = "Cases",
    fill = "Gender",
    title = "Patient Summary",
    subtitle = str_c("Distribution of males and females for HIV mono-infected and",
                     "\n",
                     "HIV/TB co-infected patients"),
    caption = "Data from Dawany et al. (2014)"
  ) +
  theme(legend.position = "none",
        axis.text.x = element_text(color = "black", 
                                   size = 14),
        axis.text.y = element_text(color = "black", 
                                   size = 14),
        axis.title = element_text(face="bold"),
        plot.caption = element_text(face = "italic"))

ggsave("../results/barplot_patient_group.png", plot = barplot_patient_group)
show(barplot_patient_group)
```

The distribution of males and females for each patient group can also be summarized in a table.

```{r}
#| label: 03_Summary table
dat_clean |>
  mutate(Gender = factor(gender),
         Disease = factor(disease_state)) |> 
  table1(x = formula(~ Gender | Disease))


```

From the above it is seen, that the gender of the HIV/TB co-infected patients is close to equal distributed between male and female, and that the gender of the HIV mono-infected is unequal distributed with 72% females and only 27% males.
