==> R/04_augment.qmd <==
---
title: "Gene Expression Analysis"
format:
  html:
    embed-resources: true
editor: visual
---

# Augment data

## Setup

Loading required packages.

```{r}
#| label: 04_Load packages
#| message: false
library("tidyverse")
library("preprocessCore")
```

Load the clean data.

```{r}
#| label: 04_Load data
#| message: false
dat_clean <- read_tsv(file = "../data/02_dat_clean.tsv")
```

In the following the gene expression data will be normalized and log2 transformed. In order to do so the clean data is split into gene expression data and meta data.

```{r}
#| label: 04_Split data
dat_expr <- dat_clean |> 
  dplyr::select(starts_with("ILMN")) 
  
dat_meta <- dat_clean |> 
  dplyr::select(!starts_with("ILMN"))

```

## Normalization

The purpose of normalization is to minimize technical variability such that the remaining variation can be explained by biological differences. Technical variation can occur because of preparation differences or differences in the production of the microarrays and the processing of the arrays (for instance scanner differences). A general strategy is to re-distribute signal intensities across all samples such that they now all have the same distribution in order to compare the data from multiple samples.

The gene expression levels in the cleaned data are normalized by quantile normalization implemented in the function "r normalize.quantiles()". This ensures that expression levels can be compared between patients in the data set. The algorithm is as follows:

1.  Sort the the expression levels for each patient.

2.  Calculate the mean expression level of the genes within the same rank.

3.  Assign this mean to each gene within this rank.

4.  Rearranging the genes for each patient to obtain the original order.

```{r}
#| label: 04_Normalise dat_expr

dat_expr_matrix <- as.matrix(t(dat_expr))

colnames(dat_expr_matrix) <- dat_meta |> 
  pull(identifier) 

dat_expr <- normalize.quantiles(x = dat_expr_matrix,
                                keep.names = TRUE) |>
  t() |> 
  as_tibble(rownames = NA)

```

## Log Transformation

In the following, the gene expression data is log2 transformed. This method is commonly used for microarray data to handle the issue that higher expression levels tend to have higher variance and vice versa.

```{r}
#| label: 04_Log-transformation
dat_expr <- dat_expr |> 
  rownames_to_column(var = "identifier") |> 
  mutate(across(starts_with("ILMN"), log2))
```

The normalized and log2 transformed gene expression data and the meta-data are combined.

```{r}
dat_aug <- inner_join(x = dat_meta,
                      y = dat_expr,
                      by = "identifier")
```

## Write Data to Files

The data is saved as .tsv-files.

```{r}
#| label: 04_Write tsv-files
dat_aug |> 
  write_tsv(file = "../data/04_dat_aug.tsv")
```

This was just a way to briefly see the effect and is not expected to be included in the final code:

## Effect of Normalization

To better understand the normalization, the distribution of the expression level is examined before normalization.

```{r}
non_normalized_expr <- dat_clean |> 
  pivot_longer(cols = starts_with("ILMN"),
               names_to = "gene_id",
               values_to = "expression") |> 
  mutate(log2_expression = log2(expression)) |> 
  ggplot(mapping = aes(x = geo_accession,
                       y = log2_expression)) +
  geom_boxplot(outlier.shape = NA,
               color = "royalblue",
               fill = "royalblue",
               alpha = 0.2) +
  coord_cartesian(ylim =  c(5, 11)) +
  theme_minimal(base_size = 16) + 
  labs(x = "Patient",
       y = "Expression level (log2)",
       title = "Non-normalized expression level",
       subtitle = str_c("Distribution of gene expression level for each patient",
                        "\n",
                        "without normalization (most extreme values are removed)"),
       caption = "Data from Dawany et al. (2014)") +
  theme(axis.text.x = element_blank(),
        axis.text.y = element_blank(),
        axis.title = element_text(face="bold"),
        plot.caption = element_text(face = "italic"),
        legend.text = element_text(size = 14))

ggsave("../results/non_normalized_expr.png", plot = non_normalized_expr)
show(non_normalized_expr)
```

After normalization, the quantiles for the expression level is equal for each patient.

```{r}
normalized_expr <- dat_aug |> 
  pivot_longer(cols = starts_with("ILMN"),
               names_to = "gene_id",
               values_to = "log2_norm_expression") |> 
  ggplot(mapping = aes(x = geo_accession,
                       y = log2_norm_expression)) +
  geom_boxplot(outlier.shape = NA,
               color = "royalblue",
               fill = "royalblue",
               alpha = 0.2) +
  coord_cartesian(ylim =  c(6, 11)) +
  theme_minimal(base_size = 16) + 
  labs(x = "Patient",
       y = "Expression level (log2)",
       title = "Normalized expression level",
       subtitle = str_c("Distribution of gene expression level for each patient",
                        "\n",
                        "with normalization (most extreme values are removed)"),
       caption = "Data from Dawany et al. (2014)") +
  theme(axis.text.x = element_blank(),
        axis.text.y = element_blank(),
        axis.title = element_text(face="bold"),
        plot.caption = element_text(face = "italic"),
        legend.text = element_text(size = 14))
 
ggsave("../results/normalized_expr.png", plot = normalized_expr)
show(normalized_expr)
```
