==> R/05_analysis_1.qmd <==
---
title: "Gene Expression Analysis"
format:
  html:
    embed-resources: true
editor: visual
---

# PCA analysis

## Setup

Loading required packages.

```{r}
#| label: 05_Load packages
#| message: false
library('tidyverse')
library("stringr")
library('broom')
```

Load the augment data.

```{r}
#| label: 05_Load data
#| message: false
dat_aug <- read_tsv(file = "../data/04_dat_aug.tsv")
```

## Principal Component Analysis

In the following a Principal Component Analysis (PCA) is carried out on the augment data.

```{r}
#| label: 05_PCA 
pca_fit <- dat_aug |> 
  select(starts_with("ILMN")) |>
  prcomp()

pca_fit_aug <- pca_fit |> 
  augment(dat_aug)
```

### Plot Principal Components

The Principal Components (PC) of the PCA is plotted against each other.

First, the first PC is plotted against the second PC.

```{r}
#| label: 05_Plot PC1 vs PC2
scatter_pc1_pc2 <- ggplot(data = pca_fit_aug,
                          mapping = aes(x = .fittedPC1, 
                                        y = .fittedPC2, 
                                        color = disease_state, 
                                        shape = gender)) + 
  geom_point(size = 3,
             alpha = 0.8) +
  scale_color_manual(
    values = c("HIV" = "royalblue", 
               "HIV/TB" = "tomato")) +
  theme_minimal(base_size = 16) + 
  labs(x = "PC1",
       y = "PC2",
       color = "Disease state",
       shape = "Gender",
       title = "Principal Component Analysis",
       subtitle = str_c("Each patient represented by PC coordinates",
                        "\n",
                        "colored by disease state"),
       caption = "Data from Dawany et al. (2014)") +
  theme(axis.text.x = element_text(color = "black", 
                                   size = 14),
        axis.text.y = element_text(color = "black", 
                                   size = 14),
        axis.title = element_text(face="bold"),
        plot.caption = element_text(face = "italic"),
        legend.text = element_text(size = 14))

ggsave("../results/scatter_pc1_pc2.png", plot = scatter_pc1_pc2)
show(scatter_pc1_pc2)
```

From above it is seen, that PC1 to some extend is able to distinguis between HIV/TB co-infected and HIV mono-infected patients, while PC2 show no prominent pattern related to infection or gender.

### Plot the Variance Explained

The variance explained by each PC is plotted.

```{r}
#| label: 05_Plot variance explained
pca_fit_eigen <- pca_fit |> 
  tidy(matrix = "eigenvalues")

var_explained <- ggplot(data = pca_fit_eigen,
                        mapping = aes(x = PC, 
                                      y = percent)) +
  geom_col(fill = "royalblue") +
  geom_smooth(mapping = aes(x = PC, 
                            y = cumulative), 
              method = "loess", 
              formula = y ~ x, 
              color = "royalblue",
              se = FALSE) +
  geom_hline(yintercept = 0.90) +
  annotate("text", 
           x=10, 
           y=0.85, 
           label="90% variance explained",
           size = 4) +
  scale_y_continuous(
    labels = scales::percent_format(),
    expand = expansion(mult = c(0, 0.01))) +
  theme_minimal(base_size = 16) + 
  labs(x = "PCs",
       y = "Variance explained",
       title = "Explained Variance",
       subtitle = "Variance explained by the first PCs",
       caption = "Data from Dawany et al. (2014)") +
  theme(axis.text.x = element_text(color = "black", 
                                   size = 14),
        axis.text.y = element_text(color = "black", 
                                   size = 14),
        axis.title = element_text(face="bold"),
        plot.caption = element_text(face = "italic"),
        legend.text = element_text(size = 14))

ggsave("../results/var_explained.png", plot = var_explained)
show(var_explained)
```

From the variance explained plot it is seen, that 35 PCs are required to explain 95% of the variance in the data.
